Feat: Pantalla final añadida Fixes #9

+++ src/components/menu/DropdownProfile.jsx
@@ -1,25 +1,46 @@
 import React from 'react';
 import { toast } from 'react-toastify';
 import './DropdownProfile.css';
-import { request } from '../axios_helper';
+import { request, getGame } from '../axios_helper';
 
 export default class DropdownProfile extends React.Component {
 
     handleLogout = async (event) => {
         event.preventDefault();
-        try {
-            const response = await request('POST', '/logout', {});
-            if (response.status === 200) {
-                window.location.href = '/';
-                sessionStorage.clear();
+        const gamePathRegex = /^\/game\/\d+$/;    
+        if (gamePathRegex.test(window.location.pathname)) {
+            try {
+                const response = await request('POST', `/game/${getGame()}/finish-game`);
+                if (response.status === 200) {
+                    const response2 = await request('POST', '/logout', {});
+                    if (response2.status === 200) {
+                        window.location.href = '/';
+                        sessionStorage.clear();
+                    }
+                }
+            } catch (error) {
+                if (error.code === 'ERR_NETWORK') {
+                    toast.error('Could not connect to server');
+                } else {
+                    toast.error('Something went wrong');
+                }
             }
-        } catch (error) {
-            if (error.code === 'ERR_NETWORK') {
-                toast.error('Could not connect to server');
-            } else
-                toast.error('Something went wrong');
+        } else {
+            try {
+                const response = await request('POST', '/logout', {});
+                if (response.status === 200) {
+                    window.location.href = '/';
+                    sessionStorage.clear();
+                }
+            } catch (error) {
+                if (error.code === 'ERR_NETWORK') {
+                    toast.error('Could not connect to server');
+                } else {
+                    toast.error('Something went wrong');
+                }
+            }
         }
-    };
+    };    
 
     render() {
         return (
